Present overlay on successful feedback submission
